refactor(crypto): Split sealing into its own module.
Sealing and AEAD ciphering are separate ideas - the first is
more opaque and just "give me bytes", whereas the cipher
traits are more low-level, and present the low level details
in a way that can be either used for sealing, or for JWE
operations.
